config(edge): switch WiFi hotspot to 5GHz band (802.11ac)
Changed hotspot configuration from 2.4GHz (band bg, channel 11) to 5GHz (band a, channel 149) for 802.11ac operation. This provides better performance and reduced interference in supported environments.

This will limit support for older devices not compatible with the 5GHz band (primarily smartphones and laptops released before 2013). This is an acceptable trade-off since the vast majority of users won't be on such old devices.